requirements.txt
MetaTrader5==5.0.37
pandas
numpy
pyyaml
